requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-storages==1.13.2
django-tinymce==3.6.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2022.7
